# frontend-mocks/openapi.yaml
openapi: 3.0.3
info:
  title: Anonytix MVP API
  version: 1.0.0
  description: API-Vertrag für das React-Frontend und das Spring-Boot-Backend.
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /companies/{companyId}/surveys:
    get:
      tags: [Surveys]
      summary: Umfragen auflisten
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "200":
          description: Umfragen
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SurveySummary"
    post:
      tags: [Surveys]
      summary: Umfrage aus einer Vorlage erstellen
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyRequest"
      responses:
        "201":
          description: Erstellte Umfrage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "400":
          $ref: "#/components/responses/BadRequest"
  /companies/{companyId}/surveys/{surveyId}:
    get:
      tags: [Surveys]
      summary: Umfrage inklusive Fragen laden
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Umfrage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Surveys]
      summary: Umfrage im Entwurfsstatus bearbeiten
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/SurveyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSurveyRequest"
      responses:
        "200":
          description: Aktualisierte Umfrage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "409":
          $ref: "#/components/responses/Conflict"
  /companies/{companyId}/surveys/{surveyId}/questions:
    post:
      tags: [Surveys]
      summary: Frage hinzufügen
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/SurveyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionWriteRequest"
      responses:
        "201":
          description: Erstellte Frage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /companies/{companyId}/surveys/{surveyId}/questions/{questionId}:
    patch:
      tags: [Surveys]
      summary: Frage bearbeiten
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/SurveyId"
        - $ref: "#/components/parameters/QuestionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionWriteRequest"
      responses:
        "200":
          description: Aktualisierte Frage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Surveys]
      summary: Frage deaktivieren
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/SurveyId"
        - $ref: "#/components/parameters/QuestionId"
      responses:
        "204":
          description: Frage deaktiviert
        "409":
          $ref: "#/components/responses/Conflict"
  /companies/{companyId}/surveys/{surveyId}/questions/order:
    put:
      tags: [Surveys]
      summary: Fragen neu sortieren
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/SurveyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [questionIds]
              properties:
                questionIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "204":
          description: Reihenfolge gespeichert
        "409":
          $ref: "#/components/responses/Conflict"
  /companies/{companyId}/surveys/{surveyId}/publish:
    post:
      tags: [Surveys]
      summary: Umfrage veröffentlichen
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Veröffentlichte Umfrage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  parameters:
...
    CompanyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SurveyId:
      name: surveyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QuestionId:
      name: questionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Anfrage ungültig
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Ressource nicht gefunden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: Konflikt mit dem aktuellen Status
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    SurveySummary:
      type: object
      required: [id, title, type, status, version]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        type:
          $ref: "#/components/schemas/SurveyType"
        status:
          $ref: "#/components/schemas/SurveyStatus"
        version:
          type: integer
    Survey:
      allOf:
        - $ref: "#/components/schemas/SurveySummary"
        - type: object
          properties:
            description:
              type: string
              nullable: true
            questions:
              type: array
              items:
                $ref: "#/components/schemas/Question"
    CreateSurveyRequest:
      type: object
      required: [title, type, templateKey]
      properties:
        title:
          type: string
          maxLength: 160
        description:
          type: string
          nullable: true
        type:
          $ref: "#/components/schemas/SurveyType"
        templateKey:
          type: string
          enum: [EMPLOYEE_SATISFACTION, EXIT_INTERVIEW]
    UpdateSurveyRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 160
        description:
          type: string
          nullable: true
    Question:
      type: object
      required: [id, text, type, category, required, position, options]
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        helpText:
          type: string
          nullable: true
        type:
          $ref: "#/components/schemas/QuestionType"
        category:
          type: string
        required:
          type: boolean
        position:
          type: integer
        minimumValue:
          type: integer
          nullable: true
        maximumValue:
          type: integer
          nullable: true
        maximumLength:
          type: integer
          nullable: true
        analyzeWithAi:
          type: boolean
        departmentIds:
          type: array
          items:
            type: string
            format: uuid
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
    QuestionWriteRequest:
      type: object
      required: [text, type, category, required]
      properties:
        text:
          type: string
          maxLength: 1000
        helpText:
          type: string
          nullable: true
        type:
          $ref: "#/components/schemas/QuestionType"
        category:
          type: string
        required:
          type: boolean
        minimumValue:
          type: integer
          nullable: true
        maximumValue:
          type: integer
          nullable: true
        maximumLength:
          type: integer
          nullable: true
        analyzeWithAi:
          type: boolean
        departmentIds:
          type: array
          items:
            type: string
            format: uuid
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOptionWriteRequest"
    QuestionOption:
      type: object
      required: [id, label, value, position]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        value:
          type: string
        position:
          type: integer
    QuestionOptionWriteRequest:
      type: object
      required: [label, value, position]
      properties:
        label:
          type: string
        value:
          type: string
        position:
          type: integer
...
    ApiError:
      type: object
      required: [timestamp, status, code, message, path, fieldErrors]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        path:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
    SurveyType:
      type: string
      enum: [PULSE, EXIT]
    SurveyStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
    QuestionType:
      type: string
      enum: [RATING, SINGLE_CHOICE, MULTI_CHOICE, BOOLEAN, TEXT]
...