# src/main/java/de/anonytix/survey/domain/QuestionOption.java
package de.anonytix.survey.domain;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.FetchType;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.JoinColumn;
import jakarta.persistence.ManyToOne;
import jakarta.persistence.Table;
import java.util.UUID;

@Entity
@Table(name = "question_options")
public class QuestionOption {

    @Id
    @GeneratedValue(strategy = GenerationType.UUID)
    private UUID id;

    @ManyToOne(fetch = FetchType.LAZY, optional = false)
    @JoinColumn(name = "question_id", nullable = false)
    private Question question;

    @Column(nullable = false, length = 300)
    private String label;

    @Column(nullable = false, length = 100)
    private String value;

    @Column(nullable = false)
    private int position;

    protected QuestionOption() {
    }

    public QuestionOption(Question question, String label, String value, int position) {
        this.question = question;
        this.label = label;
        this.value = value;
        this.position = position;
    }

    public UUID getId() {
        return id;
    }

    public String getLabel() {
        return label;
    }

    public String getValue() {
        return value;
    }

    public int getPosition() {
        return position;
    }
}